Resuelve la deuda técnica de schema.sql y últimos detalles de forma para la defensa
Cierra los tres hallazgos de "buenas prácticas" que dejaba abiertos la
última devolución sobre v13, con la suite de tests corriendo en verde
después de cada uno (32/32, BUILD SUCCESS):

- usuarios.telefono: long/BIGINT -> String/VARCHAR(20), admite ceros
  iniciales y prefijo internacional. Cambio de bajo riesgo, sin
  aritmética ni tests que dependieran del tipo numérico.

- pedidos.fecha_hora: String/VARCHAR -> LocalDateTime/DATETIME. La
  query de historial usaba LIKE sobre el string; se reescribió como un
  rango de día completo [00:00, 24:00) sobre el nuevo tipo fecha/hora.
  Se agregó un test (buscarHistorialFiltraPorFechaDelDia) que no
  existía antes del cambio, para cubrir ese filtro end-to-end contra
  MySQL real.

- precio/subtotal/total: Double -> BigDecimal en Producto, ItemPedido
  y Pedido (con precision=10, scale=2), incluida la aritmética real en
  PedidoService (multiply/reduce en vez de operadores primitivos) y
  los parámetros totalMin/totalMax del filtro de historial, para no
  mezclar tipos contra la columna DECIMAL. Verificado además que el
  JSON servido sigue siendo un número plano (no string), por lo que el
  frontend no necesitó cambios.

Además: reescribe el README para dejar inequívoco que la contraseña
vacía de MySQL es exclusiva de XAMPP local y nunca se usa fuera de ese
contexto; y en Informe_Tesis_v13.docx (sección 7.2), reemplaza la cita
a "la auditoría (hallazgo C15)" -sin esa fuente en la bibliografía- por
una descripción de la revisión de código sin atribuirla a un documento
inexistente en Referencias.

+++ src/main/java/com/marianovidela/integrador_final/controller/PedidoController.java
@@ -10,6 +10,7 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.client.RestTemplate;
 
+import java.math.BigDecimal;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
@@ -58,8 +59,8 @@ public class PedidoController {
             @RequestParam(defaultValue = "") String clienteNombre,
             @RequestParam(defaultValue = "") String nombreProducto,
             @RequestParam(defaultValue = "") String estado,
-            @RequestParam(required = false)  Double totalMin,
-            @RequestParam(required = false)  Double totalMax,
+            @RequestParam(required = false)  BigDecimal totalMin,
+            @RequestParam(required = false)  BigDecimal totalMax,
             @RequestParam(defaultValue = "") String fecha) {
 
         Page<Pedido> resultado = pedidoService.buscarHistorial(
